requirements.txt
opencv-python
numpy
torch
pyyaml